feat: ✨ get routes direclty without fetch
